Split out core crate
