--- api/src/models/mod.rs
@@ -1,3 +1,0 @@
-mod lang;
-
-pub use lang::*;
