+++ core/src/consts.rs
